fix : Racing printFinalWinner 분리 및 리팩토링
