+++ src/main/java/racingcar/RacingMain.java
@@ -12,6 +12,8 @@ public class RacingMain {
         System.out.println("시도할 횟수는 몇 회인가요?");
         String tryCount = sc.nextLine();
         racing.run(Integer.parseInt(tryCount));
-        racing.printFinalWinner();
+        RacingPrinter racingPrinter = new RacingPrinter(racing.getList());
+        racingPrinter.chooseFinalWinners();
+        racingPrinter.printFinalWinner();
     }
 }
